schedwork thingy

+++ schedwork.cpp
@@ -21,7 +21,7 @@ static const Worker_T INVALID_ID = (unsigned int)-1;
 
 
 // Add prototypes for any helper functions here
-void scheduleHelper(const AvailabilityMatrix& avail, const size_t dailyNeed, const size_t maxShifts, DailySchedule& sched, std::vector<int>& workerShifts, int day, int col);
+void scheduleHelper(const AvailabilityMatrix& avail, const size_t dailyNeed, const size_t maxShifts, DailySchedule& sched, std::vector<int>& workerShifts, size_t day, size_t col);
 
 // Add your implementation of schedule() and other helper functions here
 
@@ -41,7 +41,7 @@ bool schedule(
     std::vector<int> workerShifts(avail[0].size(), 0);
     //reintialize sched after the sched.clear
     size_t temp = avail.size();
-    for (int i = 0; i < temp; i++){
+    for (size_t i = 0; i < temp; i++){
         std::vector<Worker_T> tempRow(dailyNeed, INVALID_ID);
         sched.push_back(tempRow);
     }
@@ -50,7 +50,7 @@ bool schedule(
 
 
 }
-void scheduleHelper(const AvailabilityMatrix& avail, const size_t dailyNeed, const size_t maxShifts, DailySchedule& sched, std::vector<int>& workerShifts, int day, int col){
+void scheduleHelper(const AvailabilityMatrix& avail, const size_t dailyNeed, const size_t maxShifts, DailySchedule& sched, std::vector<int>& workerShifts, size_t day, size_t col){
     //move to the next day if the col is out of bounds
     if (col >= sched[day].size()){
         scheduleHelper(avail, dailyNeed, maxShifts, sched, workerShifts, day + 1, 0);
